+++ simulstreaming/whisper/simul_whisper/whisper/assets/mel_filters.npz
