+++ simulstreaming/whisper/simul_whisper/whisper/normalizers/english.py
