+++ simulstreaming/whisper/whisper_streaming/whisper_online_main.py
@@ -68,7 +68,7 @@ def asr_factory(args, factory=None):
 
     # Create the OnlineASRProcessor
     if args.vac:
-        from whisper_streaming.vac_online_processor import VACOnlineASRProcessor
+        from .vac_online_processor import VACOnlineASRProcessor
         online = VACOnlineASRProcessor(args.min_chunk_size, online)
 
     if args.task == "translate":
