+++ simulstreaming/whisper/whisper_streaming/whisper_server.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-from whisper_streaming.whisper_online_main import *
+from .whisper_online_main import *
 
 import sys
 import argparse
@@ -7,44 +7,14 @@ import os
 import logging
 import json
 import numpy as np
+import socket
 
 logger = logging.getLogger(__name__)
 
 SAMPLING_RATE = 16000
 
+from simulstreaming.utils.server_utils import Connection
 
-######### Server objects
-
-import whisper_streaming.line_packet as line_packet
-import socket
-
-class Connection:
-    '''it wraps conn object'''
-    PACKET_SIZE = 32000*5*60 # 5 minutes # was: 65536
-
-    def __init__(self, conn):
-        self.conn = conn
-        self.last_line = ""
-
-        self.conn.setblocking(True)
-
-    def send(self, line):
-        '''it doesn't send the same line twice, because it was problematic in online-text-flow-events'''
-        if line == self.last_line:
-            return
-        line_packet.send_one_line(self.conn, line)
-        self.last_line = line
-
-    def receive_lines(self):
-        in_line = line_packet.receive_lines(self.conn)
-        return in_line
-
-    def non_blocking_receive_audio(self):
-        try:
-            r = self.conn.recv(self.PACKET_SIZE)
-            return r
-        except ConnectionResetError:
-            return None
 
 import io
 import soundfile
