--- translate/line_packet.py
@@ -1,1 +1,0 @@
-../whisper_streaming/line_packet.py
